backend/pyproject.toml
[project]
name = "socrates"
version = "0.1.0"
description = "SOCrates — Know your threats."
requires-python = ">=3.11"

dependencies = [
    "faker>=24.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "anthropic>=0.23.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/socrates"]